requirements.txt
langchain-community        # Extends LangChain with community-built tools
langchain-core             # Core components for building LLM apps
streamlit                  # Creates user-friendly web interfaces 
langchain                  # Enables LLM-based workflows and integrations
python-dotenv              # Manages environment variables securely
langchain_groq             # Integrates GROQ features with LangChain
transformers               # to calculate get_token_ids.
scholarly
autogen